// src/app/models/ticket.ts
export class ticket {
  ticketId: number = 0;
  ticketNum: string = '';
  dateAvailable: string = '';
  dateFlight: string = '';
  timeDeparture: string = '';
  timeLand: string = '';
  destination: string = '';
  priceTicket: string = '';
  typeTicket: string = '';
  detailsTicket: string = '';
  airPort: string = '';
  terminal: string = '';
  isshow: boolean = false;
}